リスト1-40

+++ app/Http/Controllers/HelloController.php
@@ -35,9 +35,13 @@ class HelloController extends Controller
     }
 
     public function other($msg) {
-        Storage::disk('public')->delete('bk_' . $this->fname);
+        if (Storage::disk('public')->exists('bk_' . $this->fname)) {
+            Storage::disk('public')->delete('bk_' . $this->fname);
+        }
         Storage::disk('public')->copy($this->fname, 'bk_' . $this->fname);
-        Storage::disk('local')->delete('bk_' . $this->fname);
+        if (Storage::disk('local')->exists('bk_' . $this->fname)) {
+            Storage::disk('local')->delete('bk_' . $this->fname);
+        }
         Storage::disk('local')->move('public/bk_' . $this->fname, 'bk_' . $this->fname);
         return redirect()->route('hello');
     }
